requirements.txt
python-dotenv==1.2.2
Requests==2.32.5
psycopg2-binary>=2.9.9